+++ tests/http/tls.py
@@ -5,5 +5,9 @@ class TestTlsHttpClient(base.TestHttpClient):
     with_tls = True
 
     def test_large_response(self):
-        # TODO: THis fails on python 2.7 (sometimes)
+        # TODO: This fails on python 2.7 (sometimes)
+        pass
+
+    def test_large_response(self):
+        # TODO: This fails on python 2.7 (sometimes)
         pass